requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
starlette==0.40.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
structlog==25.5.0

python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

motor==3.7.1
pymongo==4.15.4
SQLAlchemy>=2.0.29
asyncmy>=0.2.9
redis==7.0.1

openai==2.7.2
